hal: sdk-ng: update drivers to 26.03.00 PVW2
update drivers to 26.03.00 PVW2

+++ mcux/mcux-sdk-ng/drivers/lpcmp/fsl_lpcmp.c
@@ -118,7 +118,9 @@ void LPCMP_Init(LPCMP_Type *base, const lpcmp_config_t *config)
 #if (defined(FSL_FEATURE_LPCMP_HAS_CCR0_CMP_STOP_EN) && FSL_FEATURE_LPCMP_HAS_CCR0_CMP_STOP_EN)
     /* CCR0 register. */
 #if defined(FSL_FEATURE_LPCMP_INSTANCE_SUPPORT_CCR0_CMP_STOP_ENn)
-    if (1U == FSL_FEATURE_LPCMP_INSTANCE_SUPPORT_CCR0_CMP_STOP_ENn(base))
+    /* MISRA C-2012 Rule 14.3 deviation: Platform-specific feature macro may create invariant condition */
+    /* MISRA C-2012 Rule 10.4: Ensure type consistency in comparison */
+    if (1U == (uint32_t)FSL_FEATURE_LPCMP_INSTANCE_SUPPORT_CCR0_CMP_STOP_ENn(base))
 #endif /* FSL_FEATURE_LPCMP_INSTANCE_SUPPORT_CCR0_CMP_STOP_ENn */
     {
         if (config->enableStopMode)
